refactor(velvet): repair a vacuous mount oracle and retire duplicate test coverage
The shadow filter-spacer mount-absence test re-tested the patch path
its sibling already pins; it now mounts genuinely filter-less via a
parallel host. Two ParseClassNames tests duplicated byte-for-byte in
the cache fixture that owns that contract are removed, and the enter-
swap timing fixture is folded onto the shared simulated-panel base
alongside the other Motion scheduled mechanics.
