--- Packages/com.velvet.core/Runtime/Reconciler/Tests/Editor/MotionEnterSwapTimingTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e3459561cb7bc4f8e93a699ad13224d3
